rename uniffi folder to dotnet
