+++ dotnet/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+publish = false
+name = "chia-wallet-sdk-cs"
+version = "0.33.0"
+edition = "2024"
+
+[lib]
+name = "chia_wallet_sdk"
+crate-type = ["cdylib"]
+doc = false
+test = false
+
+[dependencies]
+uniffi = { workspace = true }
+chia-sdk-bindings = { workspace = true, features = ["uniffi"] }
+bindy = { workspace = true, features = ["uniffi"] }
+bindy-macro = { workspace = true }
+
+[target.aarch64-unknown-linux-gnu.dependencies]
+openssl = { version = "0.10.73", features = ["vendored"] }
+openssl-sys = { version = "0.9.108", features = ["vendored"] }
+
+[target.aarch64-unknown-linux-musl.dependencies]
+openssl = { version = "0.10.73", features = ["vendored"] }
+openssl-sys = { version = "0.9.108", features = ["vendored"] }
+
+[target.x86_64-unknown-linux-gnu.dependencies]
+openssl = { version = "0.10.73", features = ["vendored"] }
+openssl-sys = { version = "0.9.108", features = ["vendored"] }
+
+[target.x86_64-unknown-linux-musl.dependencies]
+openssl = { version = "0.10.73", features = ["vendored"] }
+openssl-sys = { version = "0.9.108", features = ["vendored"] }
+
+[package.metadata.cargo-machete]
+ignored = ["bindy", "chia-sdk-bindings"]
+
+[lints]
+workspace = true
