+++ dotnet/build.rs
@@ -1,0 +1,4 @@
+fn main() {
+    println!("cargo::rerun-if-changed=../bindings");
+    println!("cargo::rerun-if-changed=../bindings.json");
+}
