+++ dotnet/cs/chia-wallet-sdk.code-workspace
@@ -1,0 +1,11 @@
+{
+	"folders": [
+		{
+			"path": "../.."
+		},
+		{
+			"path": "../../../chia.wallet.dotnet"
+		}
+	],
+	"settings": {}
+}
