+++ dotnet/uniffi.toml
@@ -1,0 +1,2 @@
+[bindings.csharp]
+namespace = "ChiaWalletSdk"
