--- uniffi/build.rs
@@ -1,4 +1,0 @@
-fn main() {
-    println!("cargo::rerun-if-changed=../bindings");
-    println!("cargo::rerun-if-changed=../bindings.json");
-}
