--- uniffi/cs/.gitignore
@@ -1,11 +1,0 @@
-# .NET build output
-bin/
-obj/
-
-# NuGet packages (restored by dotnet restore)
-*.user
-*.suo
-.vs/
-
-# Test results
-TestResults/
