--- uniffi/cs/chia-wallet-sdk.code-workspace
@@ -1,11 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "../.."
-		},
-		{
-			"path": "../../../chia.wallet.dotnet"
-		}
-	],
-	"settings": {}
-}
