--- uniffi/tests/ChiaWalletSdkTests.csproj
@@ -1,43 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.12.0" />
-    <PackageReference Include="xunit" Version="2.9.3" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.8.2">
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-      <PrivateAssets>all</PrivateAssets>
-    </PackageReference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="../cs/ChiaWalletSdk.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <!-- Copy the native library to the test output directory for local runs -->
-    <None Include="../../target/release/libchia_wallet_sdk.dylib"
-          Condition="Exists('../../target/release/libchia_wallet_sdk.dylib')">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="../../target/aarch64-apple-darwin/release/libchia_wallet_sdk.dylib"
-          Condition="Exists('../../target/aarch64-apple-darwin/release/libchia_wallet_sdk.dylib')">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="../../target/release/libchia_wallet_sdk.so"
-          Condition="Exists('../../target/release/libchia_wallet_sdk.so')">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="../../target/release/chia_wallet_sdk.dll"
-          Condition="Exists('../../target/release/chia_wallet_sdk.dll')">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-</Project>
