--- uniffi/uniffi.toml
@@ -1,2 +1,0 @@
-[bindings.csharp]
-namespace = "ChiaWalletSdk"
